fixes issue #880

+++ csrc/selective_scan/selective_scan.h
@@ -7,7 +7,7 @@
 ////////////////////////////////////////////////////////////////////////////////////////////////////
 
 struct SSMScanParamsBase {
-    using index_t = uint32_t;
+    using index_t = uint64_t;
 
     int batch, seqlen, n_chunks;
     index_t a_batch_stride;
